ACTION-2085: Make spinbutton subclass composite; remove children-presentational
On at least some platforms, the buttons to increment and decrement the
values of the spinbutton are exposed to assistive technologies. This
exposure is deliberate and appropriate given that each button can be
independently clicked on to change the value. While the spec does state
that authors SHOULD ensure this functionality can be accomplished via the
keyboard, this does not address devices which lack keyboard input. We do
not want to forbid this exposure when authors and/or user agents deem it
necessary.

Note that the working group plans to add some clarifying language on
behavior with respect to spinbutton children.